+++ src/Change.js
@@ -6,13 +6,13 @@ class Change {
     }
 
     toString() {
-        let t = this.title.split('\n').map(line => `  ${line}`);
+        let t = this.title.split('\n').map((line) => `  ${line}`);
         t[0] = '-' + t[0].substr(1);
 
         if (this.description) {
             t.push('');
 
-            t = t.concat(this.description.split('\n').map(line => `  ${line}`));
+            t = t.concat(this.description.split('\n').map((line) => `  ${line}`));
         }
 
         return t.join('\n').trim();
@@ -21,7 +21,7 @@ class Change {
 
 module.exports = Change;
 
-Change.extractIssues = function(text, issues) {
+Change.extractIssues = function (text, issues) {
     return text
         .replace(/(^|[^\\])\[#(\d+)\](?=[^\(]|$)/g, (matches, start, index) => {
             if (!issues.includes(index)) {