+++ src/index.js
@@ -2,5 +2,5 @@ module.exports = {
     parser: require('./parser'),
     Change: require('./Change'),
     Changelog: require('./Changelog'),
-    Release: require('./Release')
+    Release: require('./Release'),
 };