+++ CHANGELOG.md
@@ -2,8 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [0.10.1] - 2020-05-04
+### Added
+- New command --latest-release added to cli
+- New command --release added to cli
+
+### Fixed
+- Removed trailing space in the default description generated with --init
+- Updated dependencies
 
 ## [0.10.0] - 2020-02-08
 ### Added
@@ -148,6 +157,7 @@ First version
 [#12]: https://github.com/oscarotero/keep-a-changelog/issues/12
 [#13]: https://github.com/oscarotero/keep-a-changelog/issues/13
 
+[0.10.1]: https://github.com/oscarotero/keep-a-changelog/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.9.1...v0.10.0
 [0.9.1]: https://github.com/oscarotero/keep-a-changelog/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.8.2...v0.9.0